requirements.txt
asgiref==3.3.4
Django==3.2.4
django-ckeditor==6.1.0
django-debug-toolbar==3.2.1
django-js-asset==1.2.2
loguru==0.5.3
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1

requests~=2.25.1